+++ app/controllers/factory_bot/instrumentation/application_controller.rb
@@ -36,6 +36,53 @@ module FactoryBot
             )
         end
       end
+
+      # Unfortunately +Rails.configuration.instrumentation+ is only read once
+      # and do not hot-reload on changes. Thats why we read this file manually
+      # to get always a fresh state.
+      #
+      # @return [Hash{String => Mixed}] the instrumentation scenarios
+      def instrumentation
+        config_file = FactoryBot::Instrumentation.configuration.config_file.to_s
+        template = ERB.new(Pathname.new(config_file).read)
+        load_method = YAML.respond_to?(:unsafe_load) ? :unsafe_load : :load
+        YAML.send(load_method, template.result(binding))[Rails.env] || {}
+      end
+
+      # Map all the instrumentation scenarios into groups and pass them back.
+      #
+      # @return [Hash{String => Array}] the grouped scenarios
+      def scenarios
+        res = instrumentation['scenarios'] || []
+        res.each_with_object({}) do |scenario, memo|
+          group = scenario_group(scenario['name'])
+          scenario['group'] = group
+          memo[group] = [] unless memo.key? group
+          memo[group] << scenario
+        end
+      end
+
+      # Map all the configured scenario groups to a useable hash.
+      #
+      # @return [Hash{Regexp => String}] the group mapping
+      def groups
+        (instrumentation['groups'] || {}).transform_keys do |key|
+          Regexp.new(Regexp.quote(key))
+        end
+      end
+
+      # Fetch the group name for a given scenario name. This will utilize the
+      # +SCENARIO_GROUPS+ map.
+      #
+      # @param name [String] the scenario name
+      # @return [String] the group name
+      def scenario_group(name)
+        groups.map do |pattern, group_name|
+          next unless pattern.match? name
+
+          group_name
+        end.compact.first || 'Various'
+      end
     end
   end
 end